+++ src/html/symbols/type_alias.rs
@@ -1,4 +1,4 @@
-use crate::html::jsdoc::{DocEntryCtx, SectionContentCtx, SectionCtx};
+use crate::html::render_context::RenderContext;
 use crate::html::types::render_type_def;
 use crate::html::util::*;
